// message-service/types/broker.type.ts
import { MessageType,TOPIC_TYPE, AllEvents } from "./subscription.type";

export interface PublishType {
  headers: Record<string, any>;
  topic: TOPIC_TYPE;
  event: AllEvents;
  message: Record<string, any>;
}

export type MessageHandler = (message: MessageType) => void;
export type MessageBrokerType = {
  // producer
  connectProducer: <T>() => Promise<T>;
  disconnectProducer: () => Promise<void>;
  publish: (data: PublishType) => Promise<boolean>;

  // consumer
  connectConsumer: <T>() => Promise<T>;
  disconnectConsumer: () => Promise<void>;
  subscribe: ( 
    topics: TOPIC_TYPE[],
    messageHandler: MessageHandler
  ) => Promise<void>;
};